Rename flexml_timeline to ml_timeline
