+++ src/runtime_src/xdp/profile/plugin/ml_timeline/clientDev/ml_timeline.h
@@ -14,16 +14,16 @@
  * under the License.
  */
 
-#ifndef XDP_PLUGIN_FLEXML_TIMELINE_CLIENTDEV_IMPL_H
-#define XDP_PLUGIN_FLEXML_TIMELINE_CLIENTDEV_IMPL_H
+#ifndef XDP_PLUGIN_ML_TIMELINE_CLIENTDEV_IMPL_H
+#define XDP_PLUGIN_ML_TIMELINE_CLIENTDEV_IMPL_H
 
 #include "core/include/xrt/xrt_kernel.h"
 #include "core/include/xrt/xrt_bo.h"
 
 #include "xdp/config.h"
-#include "xdp/profile/plugin/flexml_timeline/aie_config_metadata.h"
-#include "xdp/profile/plugin/flexml_timeline/flexml_timeline_impl.h"
-#include "xdp/profile/plugin/flexml_timeline/clientDev/op/op_types.h"
+#include "xdp/profile/plugin/ml_timeline/aie_config_metadata.h"
+#include "xdp/profile/plugin/ml_timeline/ml_timeline_impl.h"
+#include "xdp/profile/plugin/ml_timeline/clientDev/op/op_types.h"
 
 extern "C" {
 #include <xaiengine.h>
@@ -31,12 +31,12 @@ extern "C" {
 
 namespace xdp {
 
-  class FlexMLTimelineClientDevImpl : public FlexMLTimelineImpl
+  class MLTimelineClientDevImpl : public MLTimelineImpl
   {
     public :
-      XDP_EXPORT FlexMLTimelineClientDevImpl(VPDatabase* dB, std::shared_ptr<AieConfigMetadata> aieData);
+      XDP_EXPORT MLTimelineClientDevImpl(VPDatabase* dB, std::shared_ptr<AieConfigMetadata> aieData);
 
-      ~FlexMLTimelineClientDevImpl() = default;
+      ~MLTimelineClientDevImpl() = default;
 
       XDP_EXPORT virtual void updateAIEDevice(void* handle);
       XDP_EXPORT virtual void flushAIEDevice(void* handle);
